breakup device module into several files
