+++ belay/device_meta.py
@@ -1,0 +1,5 @@
+from autoregistry import RegistryMeta
+
+
+class DeviceMeta(RegistryMeta):
+    pass
